third_party/WebKit/LayoutTests/fast/forms/selection-wrongtype-expected.txt

Tests to ensure selectionStart, selectionEnd, selectionDirection and selectionRange throw exceptions when the input element is not a text field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS test.selectionStart threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionStart = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionEnd = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionDirection = 0 threw exception Error: InvalidStateError: DOM Exception 11.
PASS test.selectionRange() threw exception TypeError: Object #<HTMLInputElement> has no method 'selectionRange'.
PASS successfullyParsed is true

TEST COMPLETE
